requirements.txt
fastapi~=0.116.1
uvicorn
httpx~=0.28.1
python-dotenv~=1.1.1
supabase~=2.18.1